refactor: move ChangeScope into its own partial file
