+++ src/Runtime/BinaryStorage.ChangeScope.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c173405f175cf7842aefe922d077446f
